+++ src/ts-tooling.ts
@@ -26,6 +26,7 @@ export {IDisposable} from './pattern/dispose/disposable';
 export {create, createWithFactory} from './pattern/construct';
 export {EventHandler} from './pattern/events/event.handler';
 export {ReactiveStore} from './pattern/store/reactive.store';
+export {BackgroundWorker} from './pattern/background.worker/background.worker';
 
 // Utils
 export {LZCompression} from './compression/lz';